refactor: move sgid stuff out of app.ts
