+++ backend2/src/routes/index.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import sgidRouter from "./sgid";
+
+const apiRouter = Router();
+
+apiRouter.use("/", sgidRouter);
+
+export default apiRouter;
